+++ mcp_tools/smb_enum/enum4linux.py
@@ -1,0 +1,28 @@
+# mcp_tools/smb_enum/enum4linux.py
+
+from typing import Dict, Any
+
+def register_enum4linux_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def enum4linux_scan(target: str, additional_args: str = "-a") -> Dict[str, Any]:
+        """
+        Execute Enum4linux for SMB enumeration with enhanced logging.
+
+        Args:
+            target: The target IP address
+            additional_args: Additional Enum4linux arguments
+
+        Returns:
+            SMB enumeration results
+        """
+        data = {
+            "target": target,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Enum4linux: {target}")
+        result = hexstrike_client.safe_post("api/tools/enum4linux", data)
+        if result.get("success"):
+            logger.info(f"✅ Enum4linux completed for {target}")
+        else:
+            logger.error(f"❌ Enum4linux failed for {target}")
+        return result
