+++ mcp_tools/web_scan/wpscan.py
@@ -1,0 +1,28 @@
+# mcp_tools/web_scan/wpscan.py
+
+from typing import Dict, Any
+
+def register_wpscan_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def wpscan_analyze(url: str, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute WPScan for WordPress vulnerability scanning with enhanced logging.
+
+        Args:
+            url: The WordPress site URL
+            additional_args: Additional WPScan arguments
+
+        Returns:
+            WordPress vulnerability scan results
+        """
+        data = {
+            "url": url,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting WPScan: {url}")
+        result = hexstrike_client.safe_post("api/tools/wpscan", data)
+        if result.get("success"):
+            logger.info(f"✅ WPScan completed for {url}")
+        else:
+            logger.error(f"❌ WPScan failed for {url}")
+        return result
